Fix submit hacks
