+++ lib/livesecret_web/live/page_live.html.heex
@@ -22,7 +22,7 @@
         <PageComponents.decrypt_modal
           :if={@special_action == :decrypting and (@self_burned? or not is_nil(@secret.content))}
           secret={@secret}
-          changeset={Secret.changeset(@secret, %{})}
+          changeset={Secret.receiver_changeset(@secret, %{})}
         />
 
         <%= case @live_action do %>
